refactor(svelte): extract normalizeInteractionConfig from interaction types
Move capability normalization into normalize-interaction-config.ts and keep
interaction.ts as the public types + re-export surface. Co-locate
normalization characterization tests.
